% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_dramms.R
\name{download_dramms}
\alias{download_dramms}
\title{Download DRAMMS Source}
\usage{
download_dramms(version = "current", ...)
}
\arguments{
\item{version}{Version of DRAMMS to download}

\item{...}{arguments to pass to \code{\link{download.file}}}
}
\value{
Filename of downloaded file
}
\description{
Download the source .tar.gz of DRAMMS
}

