% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dramms_cmd_maker.R
\name{dramms_cmd_maker}
\alias{dramms_cmd_maker}
\title{Create DRAMMS command}
\usage{
dramms_cmd_maker(cmd = NULL, args = NULL)
}
\arguments{
\item{cmd}{Command to be executed}

\item{args}{Arguments to be added to the command.}
}
\value{
character string of command
}
\description{
Create header for DRAMMS command
}
