% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dramms_combine.R
\name{dramms_combine}
\alias{dramms_combine}
\title{Combine DRAMMS Deformation}
\usage{
dramms_combine(
  indef1,
  indef2,
  outfile = NULL,
  retimg = FALSE,
  operation = c("concatenate", "add", "subtract", "mean")
)
}
\arguments{
\item{indef1}{Input Deformation Field 1}

\item{indef2}{Input Deformation Field 2}

\item{outfile}{output deformation field fielname}

\item{retimg}{return nifti object versus output image}

\item{operation}{Way to combine the deformations}
}
\value{
If \code{retimg = TRUE}, a \code{nifti} object.  
If \code{retimg = FALSE}, a the filename of the output image.
}
\description{
Combines DRAMMS deformations
}
