% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dramms_convert.R
\name{dramms_convert}
\alias{dramms_convert}
\title{Run DRAMMS Conversion}
\usage{
dramms_convert(indef, outfile = NULL, retimg = FALSE, field_dim = NULL,
  field_pixdim = NULL, informat = c("DRAMMS", "ITK", "FSL"),
  outformat = c("DRAMMS", "ITK", "FSL"), outtype = c("field", "magnitude",
  "x", "y", "z"), change_3 = FALSE)
}
\arguments{
\item{indef}{input deformation field}

\item{outfile}{output deformation field fielname}

\item{retimg}{return nifti object versus output image}

\item{field_dim}{Dimensions of input deformation field or nifti filename to
get dimensions}

\item{field_pixdim}{Pixel dimensions of input deformation field
or nifti filename to get pixel dimensions}

\item{informat}{format of input deformation field}

\item{outformat}{format of output deformation field}

\item{outtype}{type of output, either deformation field, magnitude of
vectors, or x/y/z displacements}

\item{change_3}{Change deformation field vectors to three-dimensional
displacements}
}
\value{
If \code{retimg = TRUE}, a \code{nifti} object.
If \code{retimg = FALSE}, a the filename of the output image.
}
\description{
Convert DRAMMS deformation to new output
}

