% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dramms_defop_coord.R
\name{dramms_defop_coord}
\alias{dramms_defop_coord}
\title{Run DRAMMS Deformation Coordinate}
\usage{
dramms_defop_coord(def = NULL, coord = NULL)
}
\arguments{
\item{def}{deformation field}

\item{coord}{Character or vector to be pasted of coordinates}
}
\value{
A character of the result of the system command.
}
\description{
Runs dramms_defop mapping of coordinates
}

