% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dramms_defop_coord.R, R/dramms_defop_invert.R
\name{dramms_defop_invert}
\alias{dramms_defop_invert}
\title{Run DRAMMS Deformation Coordinate}
\usage{
dramms_defop_invert(def = NULL, outimg = NULL, retimg = FALSE)

dramms_defop_invert(def = NULL, outimg = NULL, retimg = FALSE)
}
\arguments{
\item{def}{deformation field}

\item{outimg}{Output filename}

\item{retimg}{return nifti object versus output image}

\item{coord}{Character or vector to be pasted of coordinates}

\item{def}{deformation field}
}
\value{
A character of the result of the system command.

If \code{retimg = TRUE}, a \code{nifti} object.
If \code{retimg = FALSE}, a the filename of the output image.
}
\description{
Runs dramms_defop mapping of coordinates

Runs dramms_defop with inversion on deformation
}

