% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dramms_defop_operation.R
\name{dramms_defop_operation}
\alias{dramms_defop_operation}
\title{Run DRAMMS Multiplication or Division}
\usage{
dramms_defop_operation(def = NULL, factor = 1, operation = c("multiply",
  "divide"), outimg = NULL, retimg = FALSE)
}
\arguments{
\item{def}{deformation field}

\item{factor}{float to multiply or divide}

\item{operation}{Operation to perform: multiply or divide}

\item{outimg}{Output filename}

\item{retimg}{return nifti object versus output image}
}
\value{
If \code{retimg = TRUE}, a \code{nifti} object.
If \code{retimg = FALSE}, a the filename of the output image.
}
\description{
Runs dramms_defop with multiplication or division
}

