% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dramms_jacobian.R
\name{dramms_jacobian}
\alias{dramms_jacobian}
\title{Calculate Jacobian Image}
\usage{
dramms_jacobian(def, outimg = NULL, retimg = FALSE)
}
\arguments{
\item{def}{input deformation field}

\item{outimg}{Output deformation filename}

\item{retimg}{return nifti object versus output image}
}
\value{
If \code{retimg = TRUE}, a \code{nifti} object.
If \code{retimg = FALSE}, a the filename of the output image.
}
\description{
Runs DRAMMS-jacobian on deformation image to determine jacobian
}

