% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dramms_no_affine.R
\name{dramms_no_affine}
\alias{dramms_no_affine}
\title{DRAMMS without Affine}
\usage{
dramms_no_affine(..., opts = NULL)
}
\arguments{
\item{...}{arguments to pass to \code{\link{dramms}}}

\item{opts}{Extra arguments to pass to \code{dramms}, \code{-a 0} will be added}
}
\value{
Output from \code{\link{dramms}}

If \code{retimg = TRUE}, a \code{nifti} object.  
If \code{retimg = FALSE}, a the filename of the output image.
}
\description{
Runs DRAMMS on source and target without Affine transformation
}
