% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dramms_ravens.R
\name{dramms_ravens}
\alias{dramms_ravens}
\title{Run DRAMMS RAVENS Map}
\usage{
dramms_ravens(label, labs = NULL, template, def, prefix,
  scale_factor = NULL, verbose = TRUE)
}
\arguments{
\item{label}{label image in subject space. 
Note that the label image must be in uint8 (aka byte, or uchar) datatype.}

\item{labs}{vector of values of the labels in an image.  If not 
specified, will use all labels from \code{label} image that are positive
and non-zero.  Will fail if more than 5.}

\item{template}{template image, where RAVENS map will reside.}

\item{def}{input deformation field}

\item{prefix}{prefix for all RAVENS maps.}

\item{scale_factor}{scale factor (default: 1000)}

\item{verbose}{(logical) print out command before running}
}
\value{
Result of \code{\link{system}} command
}
\description{
Runs DRAMMS Warping on source and target
}
