% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dramms_warp.R
\name{dramms_warp}
\alias{dramms_warp}
\title{Run DRAMMS Warping}
\usage{
dramms_warp(source, def = NULL, outimg = NULL,
  interpolation = c("trilinear", "nearest.neighbor"), retimg = FALSE)
}
\arguments{
\item{source}{Filename (or nifti) to match to target}

\item{def}{input deformation field}

\item{outimg}{Output filename}

\item{retimg}{return nifti object versus output image}
}
\value{
If \code{retimg = TRUE}, a \code{nifti} object.
If \code{retimg = FALSE}, a the filename of the output image.
}
\description{
Runs DRAMMS Warping on source and target
}

