% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_dramms.R
\name{install_dramms}
\alias{install_dramms}
\title{Install DRAMMS Source}
\usage{
install_dramms(tarfile = NULL, install_dir = "/usr/local/dramms")
}
\arguments{
\item{tarfile}{Name ofDRAMMS source file}

\item{install_dir}{where to put}
}
\description{
Install the source .tar.gz of DRAMMS
}
