% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dramms_warp.R
\name{dramms_warp}
\alias{dramms_warp}
\title{Run DRAMMS Warping}
\usage{
dramms_warp(source, def = NULL, outfile = NULL,
  interpolation = c("trilinear", "nearest.neighbor"), template = NULL,
  retimg = FALSE, verbose = TRUE)
}
\arguments{
\item{source}{Filename (or nifti) to match to target}

\item{def}{input deformation field}

\item{outfile}{Output filename}

\item{interpolation}{Trilinear or nearest-neighbor interpolation for warping}

\item{template}{Template used for argument -t if an affine transformed required}

\item{retimg}{return nifti object versus output image}

\item{verbose}{(logical) print out command before running}
}
\value{
If \code{retimg = TRUE}, a \code{nifti} object.  
If \code{retimg = FALSE}, a the filename of the output image.
}
\description{
Runs DRAMMS Warping on source and target
}
\note{
As DRAMMS use FLIRT to do affine transformations, please use
\code{\link{flirt}} to apply transformation of affine .mat file.
}
