% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dramms_with_ravens.R
\name{dramms_with_ravens}
\alias{dramms_with_ravens}
\title{Run DRAMMS with RAVENS option}
\usage{
dramms_with_ravens(
  source,
  target,
  outfile = NULL,
  outdef = NULL,
  retimg = FALSE,
  opts = NULL,
  label,
  labs = NULL,
  ravens_prefix = NULL,
  verbose = TRUE
)
}
\arguments{
\item{source}{Filename (or nifti) to match to target}

\item{target}{Filename (or nifti) to match source to to target}

\item{outfile}{Output filename}

\item{outdef}{output deformation field (it will be in template to subject space)}

\item{retimg}{return nifti object versus output image}

\item{opts}{Extra arguments to pass to \code{dramms}}

\item{label}{label image in subject space. 
Note that the label image must be in uint8 (aka byte, or uchar) datatype.}

\item{labs}{vector of values of the labels in an image.  If not 
specified, will use all labels from \code{label} image that are positive
and non-zero.  Will fail if more than 5.}

\item{ravens_prefix}{prefix for all RAVENS maps.}

\item{verbose}{(logical) print out command before running}
}
\value{
If \code{retimg = TRUE}, a list of \code{nifti} objects consisting
of the RAVENS images, the output file (template to subject space), and 
output deformation field (template to subject space).
If \code{retimg = FALSE}, a list of the filenames of the output image.
}
\description{
Runs DRAMMS on source and target
}
\examples{
\dontrun{
require(fslr)
temp = file.path(fsldir(), "data", "standard", "MNI152_T1_1mm.nii.gz")
# get file
#  https://github.com/muschellij2/FSLR_Data/raw/master/01-Baseline_T1.nii.gz
brain = file.path(fsldir(), "data", "standard", "MNI152_T1_1mm_brain.nii.gz")
res = dramms(source = brain, target = temp, retimg = TRUE)
}
}
